Add files via upload
PostgreSQL - Project 3.sql  is the beginning of SQL scripts for analyzing the trade data being fetched from 'yfinance'.
Project3-crypto.ipynb - Is an updated copy of the Jupyter Notebook that includes a new cell for pulling an entire block of data to operate on, rather waiting for repeated timed calls to execute over a period of hours. It has parameters for period and interval, where period is the timeframe it will capture and interval is the time between data entries.